+++ src/XRoadLib/Handler/ServiceDescriptionHandlerBase.cs
@@ -1,5 +1,6 @@
 ﻿using System.Reflection;
 using System.Web;
+using XRoadLib.Description;
 using XRoadLib.Extensions;
 using XRoadLib.Header;