+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+#### 1.0.0-alpha21 - Unreleased
+* Move producer definition to Description namespace.
+* Fix handling of max supported operation version.
+* Fix operation types in service description.
+
 #### 1.0.0-alpha20 - January 26 2016
 * Improve protocol v4.0 generated service description.
 * Bug fix: add X-Road namespace to imports when title element is used in schema.